refactor(profiles): extract top-level profiles core module
Relocate agent profiles out of the agent/ domain into a dedicated
top-level src/openhuman/profiles/ module following the canonical module
shape (types/store/paths/prompt_section/schemas/mod). RPC moves to the
'profiles' namespace (openhuman.profiles_list / profile_{select,upsert,
delete}); data file (agent_profiles.json) and struct name unchanged.

Also folds in the new per-profile selection fields on AgentProfile:
memory_sources, include_agent_conversations (default true), allowed_skills,
allowed_mcp_servers — all serde-optional/back-compat, normalised via a
shared allowlist helper (empty -> None = 'all'). Enforcement wiring lands
in a follow-up commit.

Updated all importers (web channel, factory, task_dispatcher,
delegate_to_personality) and the relocated integration tests.
