+++ src/openhuman/agent/mod.rs
@@ -29,9 +29,7 @@ pub mod host_runtime;
 pub mod library;
 pub mod memory_loader;
 pub mod multimodal;
-pub mod personality_paths;
 pub mod pformat;
-pub mod profiles;
 pub mod progress;
 /// Prompt plumbing — types, section builders, and
 /// [`SystemPromptBuilder`](prompts::SystemPromptBuilder). Moved from
