+++ src/openhuman/agent/schemas.rs
@@ -4,7 +4,6 @@ use serde_json::{Map, Value};
 
 use crate::core::all::{ControllerFuture, RegisteredController};
 use crate::core::{ControllerSchema, FieldSchema, TypeSchema};
-use crate::openhuman::agent::profiles::{AgentProfile, AgentProfileStore};
 use crate::openhuman::config::rpc as config_rpc;
 use crate::rpc::RpcOutcome;
 
@@ -24,10 +23,6 @@ pub fn all_controller_schemas() -> Vec<ControllerSchema> {
         schemas("get_definition"),
         schemas("reload_definitions"),
         schemas("triage_evaluate"),
-        schemas("profiles_list"),
-        schemas("profile_select"),
-        schemas("profile_upsert"),
-        schemas("profile_delete"),
     ]
 }
 
@@ -60,22 +55,6 @@ pub fn all_registered_controllers() -> Vec<RegisteredController> {
         RegisteredController {
             schema: schemas("triage_evaluate"),
             handler: handle_triage_evaluate,
-        },
-        RegisteredController {
-            schema: schemas("profiles_list"),
-            handler: handle_profiles_list,
-        },
-        RegisteredController {
-            schema: schemas("profile_select"),
-            handler: handle_profile_select,
-        },
-        RegisteredController {
-            schema: schemas("profile_upsert"),
-            handler: handle_profile_upsert,
-        },
-        RegisteredController {
-            schema: schemas("profile_delete"),
-            handler: handle_profile_delete,
         },
     ]
 }
@@ -166,48 +145,6 @@ pub fn schemas(function: &str) -> ControllerSchema {
             ],
             outputs: vec![json_output("result", "Triage evaluation result.")],
         },
-        "profiles_list" => ControllerSchema {
-            namespace: "agent",
-            function: "profiles_list",
-            description: "List persistent agent profiles and the active profile id.",
-            inputs: vec![],
-            outputs: vec![json_output("profiles", "Agent profile state payload.")],
-        },
-        "profile_select" => ControllerSchema {
-            namespace: "agent",
-            function: "profile_select",
-            description: "Select the active persistent agent profile.",
-            inputs: vec![required_string("profile_id", "Agent profile id.")],
-            outputs: vec![json_output(
-                "profiles",
-                "Updated agent profile state payload.",
-            )],
-        },
-        "profile_upsert" => ControllerSchema {
-            namespace: "agent",
-            function: "profile_upsert",
-            description: "Create or update an agent profile.",
-            inputs: vec![FieldSchema {
-                name: "profile",
-                ty: TypeSchema::Json,
-                comment: "Agent profile payload.",
-                required: true,
-            }],
-            outputs: vec![json_output(
-                "profiles",
-                "Updated agent profile state payload.",
-            )],
-        },
-        "profile_delete" => ControllerSchema {
-            namespace: "agent",
-            function: "profile_delete",
-            description: "Delete a custom agent profile.",
-            inputs: vec![required_string("profile_id", "Agent profile id.")],
-            outputs: vec![json_output(
-                "profiles",
-                "Updated agent profile state payload.",
-            )],
-        },
         _ => ControllerSchema {
             namespace: "agent",
             function: "unknown",
@@ -221,21 +158,6 @@ pub fn schemas(function: &str) -> ControllerSchema {
             }],
         },
     }
-}
-
-#[derive(Debug, Deserialize)]
-struct ProfileSelectParams {
-    profile_id: String,
-}
-
-#[derive(Debug, Deserialize)]
-struct ProfileUpsertParams {
-    profile: AgentProfile,
-}
-
-#[derive(Debug, Deserialize)]
-struct ProfileDeleteParams {
-    profile_id: String,
 }
 
 fn handle_chat(params: Map<String, Value>) -> ControllerFuture {
@@ -439,206 +361,6 @@ fn handle_triage_evaluate(params: Map<String, Value>) -> ControllerFuture {
     })
 }
 
-fn handle_profiles_list(_params: Map<String, Value>) -> ControllerFuture {
-    Box::pin(async move {
-        let request_id = format!("profiles-list-{}", uuid::Uuid::new_v4());
-        tracing::debug!(
-            request_id = %request_id,
-            "[rpc][agent][entry] profiles_list"
-        );
-        let config = config_rpc::load_config_with_timeout().await.map_err(|e| {
-            tracing::debug!(
-                request_id = %request_id,
-                error = %e,
-                "[rpc][agent][error] profiles_list load_config"
-            );
-            e
-        })?;
-        let state = AgentProfileStore::new(config.workspace_dir)
-            .load()
-            .map_err(|e| {
-                tracing::debug!(
-                    request_id = %request_id,
-                    error = %e,
-                    "[rpc][agent][error] profiles_list load_store"
-                );
-                e
-            })?;
-        tracing::debug!(
-            request_id = %request_id,
-            active_profile_id = %state.active_profile_id,
-            profile_count = state.profiles.len(),
-            "[rpc][agent][exit] profiles_list"
-        );
-        Ok(serde_json::json!({
-            "profiles": state.profiles,
-            "activeProfileId": state.active_profile_id,
-        }))
-    })
-}
-
-fn handle_profile_select(params: Map<String, Value>) -> ControllerFuture {
-    Box::pin(async move {
-        let request_id = format!("profile-select-{}", uuid::Uuid::new_v4());
-        tracing::debug!(
-            request_id = %request_id,
-            "[rpc][agent][entry] profile_select"
-        );
-        let p = deserialize_params::<ProfileSelectParams>(params)?;
-        tracing::debug!(
-            request_id = %request_id,
-            profile_id = %p.profile_id,
-            "[rpc][agent] profile_select params"
-        );
-        let config = config_rpc::load_config_with_timeout().await.map_err(|e| {
-            tracing::debug!(
-                request_id = %request_id,
-                profile_id = %p.profile_id,
-                error = %e,
-                "[rpc][agent][error] profile_select load_config"
-            );
-            e
-        })?;
-        let state = AgentProfileStore::new(config.workspace_dir)
-            .select(&p.profile_id)
-            .map_err(|e| {
-                tracing::debug!(
-                    request_id = %request_id,
-                    profile_id = %p.profile_id,
-                    error = %e,
-                    "[rpc][agent][error] profile_select store"
-                );
-                e
-            })?;
-        tracing::debug!(
-            request_id = %request_id,
-            profile_id = %p.profile_id,
-            active_profile_id = %state.active_profile_id,
-            "[rpc][agent][exit] profile_select"
-        );
-        Ok(serde_json::json!({
-            "profiles": state.profiles,
-            "activeProfileId": state.active_profile_id,
-        }))
-    })
-}
-
-fn handle_profile_upsert(params: Map<String, Value>) -> ControllerFuture {
-    Box::pin(async move {
-        let request_id = format!("profile-upsert-{}", uuid::Uuid::new_v4());
-        tracing::debug!(
-            request_id = %request_id,
-            "[rpc][agent][entry] profile_upsert"
-        );
-        let p = deserialize_params::<ProfileUpsertParams>(params)?;
-        tracing::debug!(
-            request_id = %request_id,
-            profile_id = %p.profile.id,
-            agent_id = %p.profile.agent_id,
-            "[rpc][agent] profile_upsert params"
-        );
-        if let Some(registry) = crate::openhuman::agent::harness::AgentDefinitionRegistry::global()
-        {
-            let agent_id = p.profile.agent_id.trim();
-            if !agent_id.is_empty() && registry.get(agent_id).is_none() {
-                tracing::debug!(
-                    request_id = %request_id,
-                    profile_id = %p.profile.id,
-                    agent_id,
-                    "[rpc][agent][error] profile_upsert unknown_agent"
-                );
-                return Err(format!("agent definition '{agent_id}' not found"));
-            }
-            tracing::debug!(
-                request_id = %request_id,
-                profile_id = %p.profile.id,
-                agent_id,
-                "[rpc][agent] profile_upsert registry_ok"
-            );
-        } else {
-            tracing::debug!(
-                request_id = %request_id,
-                "[rpc][agent] profile_upsert registry_unavailable"
-            );
-        }
-        let config = config_rpc::load_config_with_timeout().await.map_err(|e| {
-            tracing::debug!(
-                request_id = %request_id,
-                error = %e,
-                "[rpc][agent][error] profile_upsert load_config"
-            );
-            e
-        })?;
-        let state = AgentProfileStore::new(config.workspace_dir)
-            .upsert(p.profile)
-            .map_err(|e| {
-                tracing::debug!(
-                    request_id = %request_id,
-                    error = %e,
-                    "[rpc][agent][error] profile_upsert store"
-                );
-                e
-            })?;
-        tracing::debug!(
-            request_id = %request_id,
-            active_profile_id = %state.active_profile_id,
-            profile_count = state.profiles.len(),
-            "[rpc][agent][exit] profile_upsert"
-        );
-        Ok(serde_json::json!({
-            "profiles": state.profiles,
-            "activeProfileId": state.active_profile_id,
-        }))
-    })
-}
-
-fn handle_profile_delete(params: Map<String, Value>) -> ControllerFuture {
-    Box::pin(async move {
-        let request_id = format!("profile-delete-{}", uuid::Uuid::new_v4());
-        tracing::debug!(
-            request_id = %request_id,
-            "[rpc][agent][entry] profile_delete"
-        );
-        let p = deserialize_params::<ProfileDeleteParams>(params)?;
-        tracing::debug!(
-            request_id = %request_id,
-            profile_id = %p.profile_id,
-            "[rpc][agent] profile_delete params"
-        );
-        let config = config_rpc::load_config_with_timeout().await.map_err(|e| {
-            tracing::debug!(
-                request_id = %request_id,
-                profile_id = %p.profile_id,
-                error = %e,
-                "[rpc][agent][error] profile_delete load_config"
-            );
-            e
-        })?;
-        let state = AgentProfileStore::new(config.workspace_dir)
-            .delete(&p.profile_id)
-            .map_err(|e| {
-                tracing::debug!(
-                    request_id = %request_id,
-                    profile_id = %p.profile_id,
-                    error = %e,
-                    "[rpc][agent][error] profile_delete store"
-                );
-                e
-            })?;
-        tracing::debug!(
-            request_id = %request_id,
-            profile_id = %p.profile_id,
-            active_profile_id = %state.active_profile_id,
-            profile_count = state.profiles.len(),
-            "[rpc][agent][exit] profile_delete"
-        );
-        Ok(serde_json::json!({
-            "profiles": state.profiles,
-            "activeProfileId": state.active_profile_id,
-        }))
-    })
-}
-
 fn deserialize_params<T: DeserializeOwned>(params: Map<String, Value>) -> Result<T, String> {
     serde_json::from_value(Value::Object(params)).map_err(|e| format!("invalid params: {e}"))
 }
@@ -687,8 +409,6 @@ fn to_json<T: serde::Serialize>(outcome: RpcOutcome<T>) -> Result<Value, String>
 mod tests {
     use super::*;
     use crate::core::TypeSchema;
-    use crate::openhuman::agent::profiles::DEFAULT_PROFILE_ID;
-    use crate::openhuman::config::TEST_ENV_LOCK as ENV_LOCK;
     use serde_json::json;
 
     #[test]
@@ -705,10 +425,6 @@ mod tests {
                 "get_definition",
                 "reload_definitions",
                 "triage_evaluate",
-                "profiles_list",
-                "profile_select",
-                "profile_upsert",
-                "profile_delete",
             ]
         );
         assert_eq!(schemas.len(), all_registered_controllers().len());
@@ -731,14 +447,6 @@ mod tests {
             .inputs
             .iter()
             .any(|input| input.name == "dry_run" && !input.required));
-
-        let profiles = schemas("profiles_list");
-        assert_eq!(profiles.inputs.len(), 0);
-        let profile_select = schemas("profile_select");
-        assert!(profile_select
-            .inputs
-            .iter()
-            .any(|input| input.name == "profile_id" && input.required));
 
         let unknown = schemas("nope");
         assert_eq!(unknown.function, "unknown");
@@ -816,103 +524,5 @@ mod tests {
         let value =
             to_json(RpcOutcome::new(json!({ "ok": true }), Vec::new())).expect("json outcome");
         assert_eq!(value["ok"], json!(true));
-    }
-
-    struct WorkspaceEnvGuard {
-        previous: Option<std::ffi::OsString>,
-    }
-
-    impl WorkspaceEnvGuard {
-        fn set(path: &std::path::Path) -> Self {
-            let previous = std::env::var_os("OPENHUMAN_WORKSPACE");
-            unsafe {
-                std::env::set_var("OPENHUMAN_WORKSPACE", path);
-            }
-            Self { previous }
-        }
-    }
-
-    impl Drop for WorkspaceEnvGuard {
-        fn drop(&mut self) {
-            match self.previous.take() {
-                Some(value) => unsafe {
-                    std::env::set_var("OPENHUMAN_WORKSPACE", value);
-                },
-                None => unsafe {
-                    std::env::remove_var("OPENHUMAN_WORKSPACE");
-                },
-            }
-        }
-    }
-
-    #[tokio::test]
-    async fn profile_handlers_persist_and_return_profile_state() {
-        let _guard = ENV_LOCK.lock().unwrap_or_else(|e| e.into_inner());
-        let temp = tempfile::tempdir().expect("tempdir");
-        let _env = WorkspaceEnvGuard::set(temp.path());
-
-        let upserted = handle_profile_upsert(Map::from_iter([(
-            "profile".into(),
-            json!({
-                "id": "writer",
-                "name": "Writer",
-                "description": "Draft concise copy",
-                "agentId": "orchestrator",
-                "modelOverride": "agentic-v1",
-                "temperature": 0.2,
-                "systemPromptSuffix": "Use a crisp tone.",
-                "allowedTools": ["todo"],
-                "builtIn": false,
-            }),
-        )]))
-        .await
-        .expect("profile upsert");
-        assert_eq!(upserted["activeProfileId"], DEFAULT_PROFILE_ID);
-        assert!(upserted["profiles"]
-            .as_array()
-            .expect("profiles array")
-            .iter()
-            .any(|profile| profile["id"] == "writer"));
-
-        let selected = handle_profile_select(Map::from_iter([(
-            "profile_id".into(),
-            Value::String("writer".into()),
-        )]))
-        .await
-        .expect("profile select");
-        assert_eq!(selected["activeProfileId"], "writer");
-
-        let listed = handle_profiles_list(Map::new())
-            .await
-            .expect("profiles list");
-        assert_eq!(listed["activeProfileId"], "writer");
-
-        let deleted = handle_profile_delete(Map::from_iter([(
-            "profile_id".into(),
-            Value::String("writer".into()),
-        )]))
-        .await
-        .expect("profile delete");
-        assert_eq!(deleted["activeProfileId"], DEFAULT_PROFILE_ID);
-    }
-
-    #[tokio::test]
-    async fn profile_upsert_rejects_unknown_registered_agent_id() {
-        let _guard = ENV_LOCK.lock().unwrap_or_else(|e| e.into_inner());
-        let _ = crate::openhuman::agent::harness::AgentDefinitionRegistry::init_global_builtins();
-
-        let err = handle_profile_upsert(Map::from_iter([(
-            "profile".into(),
-            json!({
-                "id": "bad",
-                "name": "Bad",
-                "description": "",
-                "agentId": "__missing_agent__",
-                "builtIn": false,
-            }),
-        )]))
-        .await
-        .expect_err("unknown agent should fail before store write");
-        assert!(err.contains("agent definition"), "err: {err}");
     }
 }
