+++ tests/personality_e2e.rs
@@ -18,14 +18,6 @@ use std::sync::Arc;
 use serde_json::json;
 use tempfile::tempdir;
 
-use openhuman_core::openhuman::agent::personality_paths::{
-    filter_integrations, memory_subdir_for_suffix, memory_tree_subdir_for_suffix,
-    resolve_personality_memory_md, resolve_personality_soul, session_raw_subdir_for_suffix,
-    HasToolkit, PersonalityContext,
-};
-use openhuman_core::openhuman::agent::profiles::{
-    built_in_profiles, AgentProfile, AgentProfileStore, DEFAULT_PROFILE_ID,
-};
 use openhuman_core::openhuman::agent::prompts::types::LearnedContextData;
 use openhuman_core::openhuman::agent::prompts::{
     IdentitySection, PersonalityRosterEntry, PersonalityRosterSection, PromptContext,
@@ -35,6 +27,14 @@ use openhuman_core::openhuman::embeddings::NoopEmbedding;
 use openhuman_core::openhuman::memory::{NamespaceDocumentInput, UnifiedMemory};
 use openhuman_core::openhuman::memory_conversations::{
     ensure_thread, list_threads, update_thread_title, ConversationStore, CreateConversationThread,
+};
+use openhuman_core::openhuman::profiles::{
+    built_in_profiles, AgentProfile, AgentProfileStore, DEFAULT_PROFILE_ID,
+};
+use openhuman_core::openhuman::profiles::{
+    filter_integrations, memory_subdir_for_suffix, memory_tree_subdir_for_suffix,
+    resolve_personality_memory_md, resolve_personality_soul, session_raw_subdir_for_suffix,
+    HasToolkit, PersonalityContext,
 };
 
 // ─────────────────────────────────────────────────────────────────────────────
@@ -57,6 +57,10 @@ fn make_profile(id: &str, name: &str) -> AgentProfile {
         soul_md: None,
         soul_md_path: None,
         composio_integrations: None,
+        memory_sources: None,
+        include_agent_conversations: true,
+        allowed_skills: None,
+        allowed_mcp_servers: None,
         memory_dir_suffix: None,
         is_master: false,
         sort_order: None,
